+++ backend/controllers/public/search/index.ts
@@ -1,31 +1,20 @@
 import { Request, Response } from "express";
-import { makeSearchOption } from "../../../utils/makePrismaObtion";
-import { getAlbumCovers } from "../../../models/albums";
-import { getArtistCovers } from "../../../models/artists";
-import { getTrackCovers } from "../../../models/tracks";
+import getAll from "./getAll";
+import getTracks from "./getAlbums";
+import getAlbums from "./getTracks";
+import getArtists from "./getArtists";
 
 interface Controller {
-  search(req: Request, res: Response): Promise<void>;
+  getAll(req: Request, res: Response): Promise<void>;
+  getTracks(req: Request, res: Response): Promise<void>;
+  getAlbums(req: Request, res: Response): Promise<void>;
+  getArtists(req: Request, res: Response): Promise<void>;
 }
 
-const search = async (req: Request, res: Response): Promise<void> => {
-  const albumFilter = makeSearchOption(req.query, "albumName");
-  const trackFilter = makeSearchOption(req.query, "trackName");
-  const artistFilter = makeSearchOption(req.query, "artistName");
-
-  try {
-    const result: any = {};
-
-    // TODO : PROMISE ALL
-    result.Albums = await getAlbumCovers(albumFilter);
-    result.Tracks = await getTrackCovers(trackFilter);
-    result.Artists = await getArtistCovers(artistFilter);
-
-    res.status(200).json(result);
-  } catch (err) {
-    res.status(500).json({ statusCode: 500, message: err.message });
-  }
+const controller: Controller = {
+  getAll,
+  getTracks,
+  getAlbums,
+  getArtists,
 };
-
-const controller: Controller = { search };
 export default controller;
