+++ backend/routes/public/search/index.ts
@@ -2,6 +2,6 @@ import { Router } from "express";
 import searchController from "../../../controllers/public/search";
 
 const router = Router();
-router.get("/", searchController.search);
+router.get("/", searchController.getAll);
 
 export default router;
